# Remove an unnecessary string allocation