This removes an potentially unneeded allocation when processing messages. Previous to this change, the processMsg function always converted the subject from a []byte to a string, thus allocating, and used it to lookup or set a cache entry.

By moving the conversion for the cache lookup into the map access, we can use an optimization introduced in Go 1.3 that avoids allocating the string for the map lookup.

In case where we insert into the cache we still need to allocate, because we need a string for the Match function and also because setting the map key still needs to allocate, if the key isn't already in the map (if I remember correctly).

These are the results from the publish benchmarks from github.com/nats-io/nats/test:

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkPublishSpeed-4            394           339           -13.96%
BenchmarkPublishSpeedViaChan-4     2005          2013          +0.40%

benchmark                          old allocs     new allocs     delta
BenchmarkPublishSpeed-4            1              0              -100.00%
BenchmarkPublishSpeedViaChan-4     4              3              -25.00%

benchmark                          old bytes     new bytes     delta
BenchmarkPublishSpeed-4            3             0             -100.00%
BenchmarkPublishSpeedViaChan-4     169           167           -1.18%
```

Both old and new number are the best of 3 runs. The slow down in BenchmarkPublishSpeedViaChan is just noise and fluctuates between + and - on the machine I used to test.